# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST multi-session para WhatsApp usando whatsmeow. Baseada na WuzAPI com campos JSON em PascalCase.",
        "title": "FioZap API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "FioZap Support",
            "url": "https://github.com/fiozap/fiozap"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/sessions/{name}/newsletters": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lista canais que a sessao segue (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "newsletters"
                ],
                "summary": "Listar canais",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria um novo canal/newsletter (nao implementado)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "newsletters"
                ],
                "summary": "Criar canal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do canal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateNewsletterRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/newsletters/{newsletterJid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna informacoes de um canal (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "newsletters"
                ],
                "summary": "Obter canal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JID do canal",
                        "name": "newsletterJid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/newsletters/{newsletterJid}/follow": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Segue um canal (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "newsletters"
                ],
                "summary": "Seguir canal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JID do canal",
                        "name": "newsletterJid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/newsletters/{newsletterJid}/mute": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ativa/desativa notificacoes do canal (nao implementado)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "newsletters"
                ],
                "summary": "Silenciar canal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JID do canal",
                        "name": "newsletterJid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Estado mute",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewsletterMuteRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/newsletters/{newsletterJid}/reaction": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Envia reacao em mensagem do canal (nao implementado)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "newsletters"
                ],
                "summary": "Reagir em mensagem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JID do canal",
                        "name": "newsletterJid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da reacao",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewsletterReactionRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/newsletters/{newsletterJid}/unfollow": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deixa de seguir um canal (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "newsletters"
                ],
                "summary": "Deixar de seguir",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JID do canal",
                        "name": "newsletterJid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/sessions/{name}/privacy": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna configuracoes de privacidade da sessao (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "privacy"
                ],
                "summary": "Obter configuracoes de privacidade",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Altera uma configuracao de privacidade (nao implementado)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "privacy"
                ],
                "summary": "Alterar privacidade",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Configuracao",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SetPrivacyRequest"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/privacy/status": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna configuracoes de privacidade do status (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "privacy"
                ],
                "summary": "Privacidade do status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateNewsletterRequest": {
            "type": "object",
            "properties": {
                "Description": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Picture": {
                    "description": "base64",
                    "type": "string"
                }
            }
        },
...
        "dto.NewsletterMuteRequest": {
            "type": "object",
            "properties": {
                "Mute": {
                    "type": "boolean"
                }
            }
        },
        "dto.NewsletterReactionRequest": {
            "type": "object",
            "properties": {
                "Reaction": {
                    "description": "emoji ou vazio para remover",
                    "type": "string"
                },
                "ServerID": {
                    "type": "string"
                }
            }
        },
...
        "dto.Response": {
            "description": "Resposta padrao da API FioZap",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "string",
                    "example": "error message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.SetPrivacyRequest": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "groupAdd, lastSeen, status, profile, readReceipts, callAdd, online",
                    "type": "string"
                },
                "Value": {
                    "description": "all, contacts, contact_blacklist, none, known, match_last_seen",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token global (GLOBAL_API_TOKEN) ou token da sessao",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}